MVC refactoring done
