+++ backend/.env
@@ -1,3 +1,4 @@
 API_URL = /api/v1
 CONNECTION_STRING = mongodb+srv://vicran:VINcent98&^@taxsystem.akw2t.mongodb.net/TaxDB?retryWrites=true&w=majority
-secret = vehicle-tax-collection-system
+JWT_SECRET = vehicle-tax-collection-system
+JWT_EXPIRY = 120d
