+++ backend/app.js
@@ -10,7 +10,7 @@ const cors = require('cors');
 require('dotenv/config');
 
 const api = process.env.API_URL;
-var authJwt = require('./helpers/jwt');
+// var authJwt = require('./helpers/jwt');
 var errorHandler = require('./helpers/error-handler');
 
 app.use(cors());
@@ -21,7 +21,7 @@ app.use(bodyParser.json());
 
 //* HTTP loggers details
 app.use(morgan('tiny'));
-app.use(authJwt());
+// app.use(authJwt());
 app.use(errorHandler);
