+++ backend/helpers/jwt.js
@@ -1,26 +1,26 @@
-const expressJwt = require('express-jwt');
+// const expressJwt = require('express-jwt');
 
-function authJwt() {
-    const secret = process.env.secret;
-    const api = process.env.API_URL;
-    return expressJwt({
-        secret,
-        algorithms: ['HS256'],
-        isRevoked: isRevoked
-    }).unless({
-        path: [
-            `${api}/users/login`,
-            `${api}/users/register`
-        ]
-    });
-}
+// function authJwt() {
+//     const secret = process.env.secret;
+//     const api = process.env.API_URL;
+//     return expressJwt({
+//         secret,
+//         algorithms: ['HS256'],
+//         isRevoked: isRevoked
+//     }).unless({ 
+//         path: [
+//             `${api}/users/login`,
+//             `${api}/users/register`
+//         ]
+//     });
+// }
 
-//* verifying the token as legitimate admin
-async function isRevoked(req, payload, done) {
-    if(!payload.isAdmin) {
-        done(null, true);
-    }
-    done();
-}
+// //* verifying the token as legitimate admin
+// async function isRevoked(req, payload, done) {
+//     if(!payload.isAdmin) {
+//         done(null, true);
+//     }
+//     done();
+// }
 
-module.exports = authJwt;
+// module.exports = authJwt;
