+++ backend/models/user.js
@@ -1,15 +1,20 @@
 const mongoose = require('mongoose');
+const validator = require('validator');
+const bcrypt = require('bcryptjs');
 
 const userSchema = mongoose.Schema({
 
     userName: {
         type: String,
-        required: true
+        required: [true, 'Please enter your name!']
     },
 
     userEmail: {
         type: String,
-        required: true
+        required: [true, 'Please provide your email!'],
+        unique: true,
+        lowercase: true,
+        validate: [validator.isEmail, 'Please enter a valid email!']
     },
 
     userDOB: {
@@ -30,8 +35,21 @@ const userSchema = mongoose.Schema({
 
     userPassword: {
         type: String,
-        required: true
+        required: [true, 'Please enter a password!'],
+        minlength: 6
     },
+
+    // userPasswordConfirm: {
+    //     type: String,
+    //     required: [true, 'Please confirm your password!'],
+    //     validate: {
+    //         //this only works on CREATE and SAVE 
+    //         validator: function(el) {
+    //             return el === this.password;
+    //         },
+    //         message: 'Passwords does not match!'
+    //     }
+    // },
 
     isAdmin: {
         type: Boolean,
@@ -48,5 +66,19 @@ userSchema.set('toJSON', {
     virtuals: true
 });
 
-exports.User = mongoose.model('User', userSchema);
-exports.userSchema = userSchema;
+userSchema.pre('save', async function(next) {
+    //only run this function if password actually gets modified
+    if(!this.isModified('userPassword'))
+        return next();
+
+    //hash the password with cost of 12
+    this.userPassword = await bcrypt.hashSync(this.userPassword, 12);
+
+    //delete the password confirm field
+    // this.userPasswordConfirm = undefined;
+
+    next();
+});
+
+const User = mongoose.model('User', userSchema);
+module.exports = User;
